Portability fixes found by turning on compiler warnings.
These fall into the following major groups:

 - Uninitialized variables. Several could definitely be left uninitialized
   in normal use, and the code was assuming they would be zero-filled.
 - Missing #includes. Minor, but found bugs.
 - Nonportable functions, notably sysinfo(), which was being called in
   several places even though it is completely different on Solaris and
   Linux. On Linux, uname() does the job.
 - Casts of types whose size can vary, mostly in printf() and sprintf()
   calls.
 - other (aliasing violations, compilation-breaking typos, broken error
   checks of e.g. fopen(), uninitialized accesses due to #if 0 in the
   wrong place).

Also remove some unused variable declarations and dead portability
functions (notably fork1()), and decorate functions that can plainly
not return with a _dt_noreturn_, newly defined in <compiler.h> to
equal __attribute__((__noreturn__)).

Finally, this actually turns on compiler warnings, and moves to
LARGEFILE64 mode (which the code is plainly meant to run in, as
it calls open64()).

A few possible bugs are just FIXMEd, because I am not yet sure if the
problematic situation can ever occur.