No longer use AT&T traditional mode for the flex lexer.
This generates very slow lexers, and relatively little-tested (because the
behaviour of AT&T lex is so irregular in some places that it is hard to
imitate). Adjust the lexer accordingly (mostly catering for the automatic
enclosure of variable expansions in parentheses).

Also fix two instances of dangerous trailing context (i.e. /-or-$, in this case
$, at the end of two |-separated parts of a rule, both of which are of
potentially variable length). This has undefined behaviour in POSIX and would
never work reliably under flex.