Remove zones, fault trapping, and process set functions; rewrite env handling.
This is really several commits in one, because the parts all rely on each other,
and disentangling them didn't seem worthwhile.

The heart of it is an attempt to move towards removal of the Solaris-specific
header <sys/procfs.h>. This is not yet accomplished, but partially: most places
that included it no longer do, and a number of helper functions (notably
prset*()) to operate on Solaris libproc are gone. Zones are gone. Fault trapping
is gone (PTRACE_TRACEME does much the same job).

The other fix entangled with this is a near-rewrite of Penv_iter(), which no
longer works by digging the environment up using the old libproc functions and
iterating through it piece by piece, but read()s it out of /proc/$pid/environ in
one go and iterates through the result. It's a dramatic simplification.

Finally, a tiny fix to our dt_ioctl() declaration: Solaris ioctl() 'val' is
'int', Linux is 'unsigned long', and the new ioctls were overflowing 'int'.