Rip out process flags, multiple-debugger support; start on dt_proc_control().
The core libproc process-control loop is still largely Solaris-specific: it
relies on echoing state down into libproc and then receiving it back rather
than (as Linux requires) tracking it itself; it relies on being able to query
libproc for process state data rather than waitpid()ding for it; and it has
an elaborate infrastructure of process-control flags which Linux simply does
not support. (Consequently, we lose some features: see INCOMPATIBILITIES.)

This change is a start at fixing that, ripping out the process/tracing flags
machinery entirely, fixing process detaching, dropping much of the machinery
relating to PS_LOST (since it seems Linux cannot 'lose control' of ptrace()d
processes, and if it does, e.g. via exec() of a privileged process, there is
surely nothing the caller can do to fix it: 'reattaching' is hopeless), and
simplifying the process control loop a bit.

Pwait() is *not* yet changed to use waitpid(), so you cannot yet use dtrace
to trace processes. But this is much closer than it was.