fd rationalization.
We still want to retain fd caching in the post-proc world because open()ing
/proc/self/{mem,stat} is about twice as slow as open() of files on normal
filesystems. But we don't need a ctlfd, we don't need to shuffle our open fds
above 256 to cater for weirdnesses of Solaris's 32-bit libc, and we don't need
to cater for people trying to use this library when /dev/std{in,out,err} are
closed.

We also don't need special abstractions for LWPs outside of core files.