Yet another round of code removal, from Psymtab.c and Pcontrol.h this time.
The focus here is to make audits easier by removing unused functions with names
similar to those that may be called with the process grabbed.  These are almost
all in Psymtab.c.  Also, we discard a bunch of ps_prochandle_t structure members
which are never used, which lets us ditch the code that maintains them.