Fix locking and signal handling problems around process termination.
With these changes, using dtrace -p to grab a process no longer hangs or
crashes.

We had several interlocking problems here.  The OpenSolaris version of dt_proc.c
had to cope with halting control threads that were blocked inside ptrace() and
libproc operations.  Neither of these are pthread cancellation points, so it
could not use thread cancellation, instead using a complex mechanism involving
low-level nonportable thread kills, explicit sends of the nonportable SIGCANCEL
signal to the dying thread, and a variable whose sole purpose was to tell the
thread that it should die.

We (obviously) cannot keep this mechanism.  In the piecemeal nonportability
conversion so far, dt_proc.c was converted to send SIGUSR1 instead of SIGCANCEL.
Unfortunately their signal dispositions differ: SIGCANCEL is defined to be
ignored, while SIGUSR1 terminates the whole process, which is not ideal, since
all we want it to do is terminate one syscall with EINTR when it is time for the
monitor thread to die.  Changing the signal disposition is possible, but ugly,
because it is process-wide.  Since we are not using ptrace() or libproc anymore,
we don't need any of this machinery and can move to using pthread cancellation
to kill the thread off, disposign of the nonportable thread kills, the
nonportable SIGCANCEL signal, and the 'please quit' flag.

However, we must take a little care here.  The cancellation cleanup function for
the monitor thread has to take the dpr_lock so as to send a broadcast on the
dpr_cv condition variable to notify interested parties (such as
dt_proc_destroy()) that it is now dead (so that dt_proc_destroy() knows that it
is safe to deallocate associated structures).  However, the condition variable
is used in other places within the control thread, so it can (but may not) be
grabbed when cancellation kicks in.  pthreads provides no mechanism for telling
whether a lock is presently grabbed, so we must manually track whether or not
the control thread has grabbed the dpr_lock, and relock it in the cleanup thread
only if it is not already locked.  (Thankfully the control thread only grabs and
releases the dpr_lock in one place.)