Support pre/postprocessing and programmable expected-failure support.
D scripts can be preprocessed through arbitrary filters, which are executable
files (generally shell scripts) named after the D script only with the suffix
.p. The results can be postprocessed through filters named .r.p: this filter is
applied before an unconditional filter which converts all hexadecimal strings
to the string {ptr}.

A test's expected failure can now be signified not just through the xfail: flag
but also via an executable named after the D script only with the suffix .x.
This should return 0 if the test is expected to succeed, and nonzero otherwise:
it can emit a line of text on stdout which is a message describing why the test
failed.