Trigger and test-execution support.
This commit adds support a @@trigger option in .d files, which if specified is
the name of a program in the test/triggers subdirectory which should be run in
parallel with DTrace and used to prod it into action.  (It should not emit
anything to stdout or stderr: if it does, it'll mess up the screen.)
Alternatively, you can specify a trigger in a .t file alongside the .d file:
this can point to triggers located in directories other than test/triggers if
you like.  The trigger can exit after its job is done, but this will not halt
the test: only DTrace exiting or timing out will do that (a timeout is
considered the same as a successful exit).

Only scripts which have an associated trigger will be executed: all other
scripts are run with -e to suppress script execution (as are all scripts when
running the regression test phase to compare the output of multiple copies of
the dtrace userspace code).

Running runtest with --no-execute will suppress execution of scripts with an
associated trigger, as well.

As usual, test failures are reported in the form of a diff in the
test/log/current/runtest.sum file, and as complete output in
test/log/current/runtest.log.  The --capture-output option will capture the
output of one or more tests which you believe to be working properly into .r
files: you probably want to specify specific tests when specifying this option,
or *every* test which doesn't fail with a nonzero exitcode and has no results
will be captured.