Fix test coverage/trigger interaction bugs.
We don't want test coverage info in triggers, because triggers are run
simultaneously with dtrace and run the risk of coverage-file overwrites,
and also because getting coverage info for several programs at once
confuses lcov into producing no useful output at all.  Test coverage
for something as small as a trigger is likely to be useless anyway.
If a trigger has significant conditionals in it, it is probably
doing something wrong.

Also, drop test coverage output into the coverage/ subdirectory of
the test log directory rather than cluttering up the top level
with it.