Do not loop over the actions list in statement disassembly.
Even though statements point within a list of actions, that list is *not* a list
of actions belonging to that statement: statements appear to contain at most one
action (though this is hard to prove, the code relies on it in other places),
and the list exists purely to assist statement destruction and ECB building.
So do not iterate over both the statement and action list, only the statement
list, to avoid catching actions that do not belong to the statement under
consideration.