Make --capture-expected: overwrite test results in some situations.
Expected-result capturing was less useful than it might have been.
If you specified --capture-expected and a test result already existed,
even if it was known to be wrong, runtest would refuse to overwrite it
on the grounds that this was a test failure.  This was intended to
prevent a run of runtest accidentally overwriting tests which are
rightly failing, but unfortunately it made it unnecessarily hard to
re-capture results if the problem was a dtrace or runtest bug causing
the expected results to be wrong.

So we soup things up so that if you specify explicit test names on the
command line, --capture-expected will overwrite their test results no
matter what, on the grounds that you probably knew what you were doing
when you asked it to concentrate on these tests alone.  If no test names
are specified, --capture-expected works as it always did.

As another convenience improvement, you can specify .r names on the
command line now too: they are automatically translated into .d.
This helps for things like

# All tests with expected results in the test/foo directory are wrong
runtest --capture-expected test/foo/*.r

(which would otherwise have to be a lot more complicated to avoid
catching tests that do not have any expected results yet.)