Syscall provider test, test/struct/rwinfo.d.
This existing test was almost acceptable, but needed adjusting to not display
the time-in-ns that the reads and writes took, as there is no way this is
reproducible.  It's associated with a new trigger that attempts to read() every
regular .d file in the same directory as the .d file using 4096-byte chunks and
write it out to a temporary file: as such the expected results will need
regeneration whenever any .d file in this directory is changed.  (Note that the
expected results will still show a different maximum and call count for read
versus write, because the read()s are done with a count of 4096, while the
write()s are done using the amount actually read(), and because one extra
read() is done per file to detect EOF.)