A variety of new/modified tests, all of which fail.
test/act/time.d: this fails because walltimestamp appears to contain the
number of nanoseconds since boot, not the number of nanoseconds since
the epoch. (It has a fairly elaborate result postprocessor to detect
cases where the walltimestamp and the output of system("date ...") are
different, and complain.)

test/script/syscall.d: this fails because it's running something using
-c, and currently it appears that things run with -c are not being
restarted at all, so dtrace waits until timeout and is silently
terminated. I added fake expected results to this existing test to make
sure that result comparison happened, otherwise runtest would have
assumed that dtrace's timing out indicated success, instead of indicating
some sort of internal deadlock.

test/vars/clause.d: This test is failing because an extra blank line is
consistently getting into the output, despite appearing nowhere in the D
file. This may be a testsuite driver bug, but if so it's a terribly
consistent one. Probably a userspace bug, but I could be wrong.