Add a mass of XFAILs.
These mostly signify unimplemented providers or things that would work if only
they had test triggers.  One indicates the unimplemented status of speculations.
(The expected results for this are faked and probably wrong, but more likely
right than not having any there at all).

We also fix test/struct/kstat.d, which is XFAILed because it depends on
Solaris-specific probes, but was still buggy because it was trying to hook up
with a trigger using $(pidof) in the @@runtest-opts line.  This is historical
cruft from before triggers existed, and would never work unless someone had
started the trigger manually first, at *just* the right instant.  We use a
trigger now, instead.