Find working D libraries while testing.
This arranges to copy those D libraries that work into a subdirectory 'dlibs',
and for the -t switch passed by the testsuite to add that subdirectory to the D
compiler's system library path (so it takes precedence over /usr/lib/dtrace).

Without this, we'll either fail to find the libraries at all, or fall over when
we try to compile a library .d file that doesn't yet work.