Look for Makefunctions, not .mailmap, to find top-of-tree.
runtest.sh needs to find the tree top to terminate its loop when hunting upwards
for a nonexistent @@option.  It used to do this by looking for .mailmap, but this
does not exist in 'git archive'd trees.  So look for Makefunctions, instead,
which must always exist, and only ever in that location.