Allow testing of installed dtrace.
Before now, you could only test against an installed dtrace by building it,
deleting the binary from the build directory and replacing it with a symlink,
then touching /usr/sbin/dtrace to make sure it was new enough.

This will not do.

So this change adds a new --use-installed option to runtest.sh, and
corresponding check-installed and check-installed-verbose targets to the
makefile, which force testing against /usr/sbin/dtrace rather than anything in
the build tree. (It still has to do some compilation, of triggers: a straight
./runtest --use-installed isn't going to work well unless the source tree has
been built, but 'make check-installed' can avoid building a dtrace that will
never be tested.)