Fix regression test support.
The regression test support was looking for everything matching dtrace-* and
assuming it should test them all.  It should restrict itself to dtraces
matching the current kernel.

Further, the check for no available dtraces was broken by the shell nullglob
option and would have always reported that dtraces were available, even when
none were, and the printout of the name of the dtrace would always print the
same thing time and time again even if you were regression-testing more than one
of them.