Don't emit the inter-loop blank line when testing.
Because this happens between the dtrace_work() loops, it appears after a line
which is pretty much random depending on how much work the kernel module has
done. This is less than ideal for testsuite runs, and introduces spurious
failures across the board.

Some test results rely on the appearance of this blank line: fix them up.