Do not change the library search path when testing an installed dtrace.
Change cf0cca5 arranged to search libdtrace/libs instead of /usr/lib/dtrace for
the DTrace standard library when testing.  This is normally a good idea, but
when testing an installed DTrace, this means you're testing a D library that
does not correspond to your DTrace, and (if make(1) has not been run) means
that finding a D library will fail, leading to spurious test failures.

So introduce a new argument to the -t option, which if 'installed' means to
bypass the library directory change, and use it when testing an installed
DTrace.