Make linking respect the syslibdir properly.  Fix testing w/o installed dtrace.
Before now, dtrace's -G option linked against /usr/lib/dtrace/drti.o, using
/usr/lib/dtrace/drti-vers as its version file: the path was derived from
_dtrace_libdir, which is set at build time via a -D option and is also used to
populate the default dtrace search path.  This is unsatisfactory because if the
user should change the search path (whether appending to it via the libdir, or
prepending via the syslibdir), that change is not respected by the -G option.
This is now fixed.

Worse, the testsuite instructs the system to look in the syslibdir so that
testing works whether or not dtrace is installed.  -G was not respecting this, so
tests of -G failed spuriously, and linked against the wrong drti.o even when
they did not fail.

Fixing this was trickier.  The testing syslibdir was libdtrace/dlibs because
that path did not change (unlike the objdir, which contains an embedded kernel
version): but this will not do if object files are to go in there as well.  So
the definition of the barely-documented -t testsuite option is changed to
accept, rather than the word 'uninstalled' or 'installed', an absolute path to
the dtrace library directory or the word 'installed', and dtrace sets the
syslibdir to the value of that option rather than hardwiring libdtrace/dlibs.

Secondly, the build system is adjusted to put the generated .d files into a
dlibs subdirectory of the objdir, and to put drti.o and drti-vers in there as
well.  We can't actually compile drti.o in there, though, because GCC drops
its dependency files in a file with suffix .d, and DTrace expects all files
with suffix .d in its libdirs to be D scripts. So we build drti.o in the
objdir, and symlink it into place.

With all this done, -G works, whether or not the testsuite is being run and
whether or not DTrace is already installed when it runs.