The raise() tests were passing inappropriately.
The tests are supposed to fire a signal with raise(), then wait for it to be
handled (detected because the trigger kills itself when it receives the signal).

Since raise() is not implemented, these should be failing, but they were passing
because they were hooking all the exit syscalls without checking the PID: so
they were detecting the termination of other processes in the system, and
quitting early.  (The expected results were also wrong, or this would have
been spotted via result comparison.)

They are all fixed up to fail properly, and marked XFAIL.