First-cut libraryization of libdtrace.
This installs the public dtrace.h header and library.

We have to build a new library at installation time, because we need to
incorporate libproc and libctf (we don't want to install a library
called libproc on a Linux system, and libctf should stay in here
until it is split out and GPLized as part of the CTF work).

It's not quite good enough because the set of exported symbols is unacceptably
large. Fixing that requires a .so conversion and library version
script (forthcoming).