Split libctf into a new repository.
This was done, and libctf relicensed under the GPLv2 and the proprietary DTrace
userspace license, to make it possible to use the DTrace libctf implementation
in the Linux kernel build process (specifically, so that dwarf2ctf can link
against it).  So libctf has migrated into a new git repository so as not to mix
up code under multiple licenses, and been renamed to libdtrace-ctf to avoid
conflicting with the existing Common Trace Format Linux library and to make it
doubly clear that the API can and will vary at the convenience of
DTrace: we will bump soname whenever we see fit.

However, we do not wish to sacrifice the ability to conveniently build (and
test) libctf along with dtrace.  libctf doesn't change very often, so git
submodules are quite well-suited for this -- but they are also annoying to use,
because everyone checking out dtrace needs to do a 'git submodule update' to get
the submodule, and people often forget.

So libdtrace-ctf is an *optional* submodule of dtrace.  If you don't do a 'git
submodule update', the libdtrace-ctf directory in the dtrace checkout remains
empty: the makefile detects a nonempty libdtrace-ctf directory and adds
libdtrace-ctf/include to the CPPFLAGS in that situation only.  Further, the two
projects share one build system, and there is a Build file at the root of the
libdtrace-ctf repository which pulls in libctf/Build and include/Build.
Both of these consult $(PROJECT) and only provide install rules if the project
is 'libdtrace-ctf'.

The net effect of all this is that if you don't initialize submodules in your
working tree, dtrace will use your installed libdtrace-ctf headers and link
against your installed libdtrace-ctf.so: if you do initialise submodules, dtrace
will build a new copy of libdtrace-ctf.so from the submodule, guaranteed to be
the right version for this checkout: if test coverage analysis is enabled, it
will extend over libdtrace-ctf as well.  However, that submodular copy of
libdtrace-ctf is never installed: you need an installed copy of
libdtrace-ctf (perhaps generated via 'make && make install' in the libdtrace-ctf
subdirectory, or via an RPM package installation) for dtrace to run once
installed.

The only failure mode now is if libdtrace-ctf grows a new interface and you have
a newer libctf in the submodule that has that interface, but an older one
installed that lacks it.  You can expect an installed dtrace to fail with
unresolved symbols in such a situation. Don't do that.  (This is never a problem
when building RPMs, only when developing.)

This involved a bit of tweaking of include paths in various places: DTrace was
including an include file in libctf that did nothing other than include
<ctf_api.h> and that was not installed: this is no longer visible in a
non-submodular build, so everything had to change to include <ctf_api.h>
directly instead.

Finally, we dropped a few preprocessor conditionals and include files that were
only needed when building libctf or which were never used at all.