Eliminate duplicate code for ELF symbol lookup.
The near-identically-named macros and types for 32-bit and 64-bit code in
<elf.h> had spawned an invasion of hundreds of lines of duplicated
near-identical code in dt_module.c. This was annoying, pointlessly repetitive,
and hid those few significant places where differing code was necessary for 32-
versus 64-bit: so, split that code out into a header file using token pasting to
produce the appropriate type and macro names, and include that header repeatedly
to generate the ELF symbol table manipulation functions.