Eliminate the concept of 'primary objects'.
This Solaris concept has no meaning in Linux.  Dropping it means we lose a bunch
of complexity from dt_module.c as well as libdtrace API to track it, and even a
linkmode option.

While we're at it, drop most mentions of the kernel's runtime loader (which has
no API in the same sense in Linux) and yet more references to tasks, zones,
projects, and contracts.

Finally, update INCOMPATIBILITIES to note some of this.  (Just because trying
to do anything with tasks, projects, zones, contracts, and primary objects
would never have worked before this change doesn't mean their absence is not
an incompatibility.)