Abstract symbol table management for DTrace.
The module management code in libdtrace's dt_module.c assumes that modules
in tDTrace are always ELF objects, and that it can therefore use libelf
to do the dirty work of mapping from symbols to addresses and vice versa.

Alas, the upcoming dt_module changes to support Linux kernel modules will
introduce a new sort of module which is not an ELF object and for which we must
implement our own address lookup.  So dt_symtab.c implements an abstract
datatype, the dt_symtab, which permits high-speed mapping from symbols to
addresses and vice versa (via a hash table and a binary-searched sorted array).
We don't need full generality for this application, but we do need to be able to
handle a large number of symbols, so the expected lifecycle of a dt_symtab is as
follows:

 - Creation: call dt_symtab_create().

 - Population: repeatedly call dt_symbol_insert().  You can also call
   dt_symbol_by_addr() at this stage, to look up the address of a symbol given
   its name.

 - Lookup: call dt_symbol_sort(), following which calls to dt_symbol_by_addr()
   are permitted.  You can insert more symbols after this, but you must call
   dt_symbol_sort() again after doing so if you wish to call dt_symbol_by_addr()
   again.

 - (optional) Packing.  This is not currently implemented, but in the end will
   hopefully permit us to compress similar components of symbol names, to avoid
   the large memory hit from dragging around tens of thousands of uncompressed
   names of kernel symbols.

The dt_symbols returned from dt_symbol_by_*() are vaguely reminiscent of ELF
symbols, possessing a name (though not implemented as a pointer into a string
table), a size, an address (like a value in ELF symbols) and an info field
exactly the same as the info field in ELF symbols, and manipulated with the same
macros.  The function dt_symbol_to_elfsym() can return a partially-populated
ELF symbol given a dt_symbol, though its st_name and st_other fields are
unpopulated, and the only thing we guarantee about st_shndx is that it is not
SHN_UNDEF.  (This is enough for DTrace's use, and a later commit will document
the visible change to libdtrace users.)