The dt_module revamp.
The old Solaris module management code is not appropriate for Linux.  Solaris
treats userspace and kernel modules almost identically: except that kernel
modules appear in a special /system/object 'objfs', the kernel modules in
memory look just like ELF objects, occupying a contiguous range of addresses
with their virtual addresses and so on corresponding to the addresses they
actually occupy in memory.

None of this is true for Linux.  Non-built-in modules do exist on disk as ELF
objects, true, but there is no API that shows them that way: even assuming you
can find the root of the tree (for which see commit
ff612336aff35ce8110df36e15765ed66115a33d), you have no assistance in mapping
module names to the names of .ko files.  Worse yet, most modules can be built in
to the kernel, whereupon they lose their separate identity and become scattered
among the thousands of functions that make up the kernel proper, and even those
modules which aren't built in are relocated and have addresses that bear no
resemblance to the addresses in the on-disk ELF objects.  So the assumptions in
DTrace that a module occupies a contiguous address range, and that some ELF
object in the filesystem can be used to find the addresses of the in-memory
modules, are also incorrect.

This revamp, together with the preceding commits, fixes all this.  Several new
abstractions are added:

 - The dt_kern_path_t.  One of these is stored in a kernpaths hash in the
   dtrace_hdl for every module on disk, and maps a module name to a full
   pathname for a single hash.

   This is populated by the new functions dt_kern_path_create() and
   dt_kern_path_destroy(): a lookup is done via dt_kern_path_lookup_by_name().
   The actual filesystem walking and module name massaging is done by
   dt_kern_path_update(), which is called (once only) by dtrace_update().
   (We use fts() for the actual walking: it may be overdesigned but it works.)

   If a kernel module is not found in this mapping, it acquires a new
   DT_DM_BUILTIN flag to indicate that it is a built-in module.  Such modules
   all share the same dt_module.dm_file, pointing at the dtrace_ctf.ko module, a
   dummy module which contains distinct CTF sections for the core kernel and for
   all built-in modules.  This is read in at most once, copied to the
   dtrace_hdl.dt_ctf_elf member, and assigned from there to each individual
   module's dm_elf member.  With the CTF section read in, CTF lookups should
   work.

With this abstraction we can do CTF lookups. But that's not enough.  We also
have to do address->name lookups in order to figure out what type we want to get
CTF information for, and sometimes name->address lookups as well.  DTrace has
functions to do all of this, but the interspersed nature of Linux built-in
modules, and their un-ELFy nature, adds complexity.

 - The dtrace_addr_range.  DTrace used to have three fields in the dt_module,
   dm_text_va, dm_data_va, and dm_bss_va, and corresponding _size fields, giving
   the address and size of the text, data and BSS sections.  This is not
   adequate for built-in kernel modules, which can have dozens of 'text' and
   'data' sections (DTrace didn't use the BSS section for anything special,
   treating it just like the data section, so I folded the two together).  So
   rather than have a single field for each, we now have an addr_range, which is
   a sorted array of (va, size)s.

 - The dt_symtab.  Introduced in b4a72221f93296816cd5eec4e7ede4e992ce116a,
   we use this in the new dt_module.dm_kernsyms to maintain a bidirectional
   mapping of name->address for every symbol in the kernel.  This mapping is
   not actually bidirectional, of course: some addresses have no associated
   symbols, and some addresses have many symbols associated with them.  In
   the latter case, one symbol is preferred over the others: dt_symtab
   implements rules similar to the existing rules that dt_module_symcomp()
   implements for userspace ELF modules.

   This mapping, like the address range mappings, is populated by
   dt_modsym_update(), which is called by dtrace_update() with every line from
   /proc/kallmodsyms in turn.  This contains one line for every symbol in the
   kernel and all loaded modules, built-in or otherwise, giving their address,
   size, type, name, and what module they're part of.  It is sorted into
   order of ascending address except that symbols in non-built-in modules
   appear in random order.  All symbols from a given non-built-in module
   will appear together, which isn't true for built-in modules.  So
   dt_modsym_update() has to parse the line it's given, find or create a
   module for it, then add the symbol to the symbol table for this
   module and extend its address ranges to cover the address of this module,
   by extending the last range if it is right on the end, or by creating
   a new range.  After this has been done for all modules, the address
   ranges and symbol tables are sorted into order.  (This is likely to lead
   to a lot of tiny ranges for non-built-in modules, since we do not coalesce
   adjacent ranges yet: an easy fix for later.)

   But there's a little extra complexity here: there are a vast number of
   symbols in the kernel which are entirely useless for DTrace and just waste
   memory.  These are skipped, except that if they would extend an existing
   address range, they still do that.  This list of superfluous symbols is
   hardwired and will surely need updating in future.

With these abstractions, the rest of the change is quite simple: turning an
address into a name or vice versa is just a matter of working over all the
modules and scanning its kernel symbol table until you find one that
matches (for address->name lookups, we look for modules with a matching address
range first).