Allow modules with no CTF data to have a one-byte dummy CTF section.
On Solaris, such modules often have a zero-byte dummy .SUNW_ctf section. I
wanted that to be true on Linux too, only to find tha thte oldish binutils in
OL6 has an objcopy that silently fails to do anything when asked to add a
section based on an empty file.  So we use a one-byte \0 file instead, and
must cater to that.