Fix the generation of errno.d.
2/3rds of the errors Linux defines were missing, specifically all errors not
defined in errno-base.h and all errors defined in terms of other errors.

Fix the first of these problems by passing '#include <errno.h>' through the
preprocessor and stripping out everything which is not a #define starting with
the letter E.

The second of these problems could be fixed while retaining the existing
sed-based mkerrno.sh, but while it is possible to implement associative arrays
in sed using regex matching on the hold space, it results in a nightmare that
only Greg Ubben could love.  So reimplement mkerrno.sh in awk, where it is
trivial to handle the errno-defined-in-terms-of-errno problem.