Compressed CTF support and shared types support.
Two things remained to be done here.  Firstly, the CTF in kernel modules is
compressed: we have to uncompress it, as libctf only knows how to uncompress CTF
that has an uncompressed header with a compressed flag bit flipped in it (and
changing that to allow uncompression of entirely compressed files would require
a disruptive ABI change).

Secondly, the dtrace_ctf module in which all shared types are stored needed
both to be loaded into a dt_module and to be used when necessary.

Loading it was a simple matter of forcing the dtrace module to depend on
dtrace_ctf, so that it is loaded into the kernel's module list with at least one
symbol showing in /proc/kallmodsyms and thus gets a dt_module generated for it.
Using it was a little more complex.  Firstly, we must add a special case for it
and note that even though it is a real kernel module, we should
*regard* it as a built-in module, so that its .SUNW_ctf.dtrace_ctf shared-types
section is consulted for CTF rather than the empty .SUNW_ctf section which holds
the types contained in the dtrace_ctf module itself (none).

Using dtrace_ctf required a slight enhancement to the existing code which
determines the relationship between pieces of loaded CTF.  The CTF file format
contains space for a parent name, but libctf unconditionally fills this with the
string 'PARENT'.  We might fix this in future, but since we only have one parent
for all modules' CTF, for now it is easy enough to state that any kernel modules
whose parent is given as "PARENT" should instead get its CTF from the dtrace_ctf
module.

This completes enough of CTF support to let you look up the types of arguments
and dive around inside structures passed as arguments.  Only dtrace_symbol_type()
and variable-type support remain.