Unknown-type-lookup reduction: struct/union/enum type names
The next stage of unknown-type-lookup reduction is to teach the lexer that when
it sees 'enum', 'struct' or 'union', the next word, if not an open brace or
< (for translators), is always a type, never an identifier.

The lexer would always come to this conclusion anyway, but it would do it by
falling all the way through id_or_type(), doing an expensive type lookup whose
results are entirely meaningless (if it's not a known type, it will be soon and
this is a type: this is also both the common case and the really expensive one:
if it is a known type, it's either a typedef forwarding of the same type or a
syntax error, and this is *still* a type).

Also add a testcase to verify that we don't regress in this area.