Support global variable lookup.
This change has three parts:

1) pull in a new libdtrace-ctf supporting maintenance of a variable->CTF type ID
   mapping
2) implement dtrace_symbol_type() in terms of that mapping for symbols in a
   kernel module
3) fix up the testsuite to use Linux types everywhere rather than Solaris types
   (for details, see ChangeLog).

We also take advantage of this to fix a couple of unclear CTF-related error
messages and add more entries to the list of uninteresting kernel symbols.

On the testsuite front we also have to adjust some expected results, both for
changes in symbol names and because we changed the wording of one error message
in libdtrace-ctf, which is given to the user in some circumstances.