Fix -xmodpath.
The new -xmodpath option was being ignored because of an initialization ordering
problem.  The module path is used during initialization of the kernpath structure,
which has to happen before the dm_path of any modules can be set.  Every module,
loaded or not, was getting its dm_path set at dtrace_update() time, which
required the kernpaths to be initialized then as well.

Unfortunately, dtrace_update() is called by dt_vopen().  DTrace options are a
per-dtrace_hdl thing, so cmd/dtrace.c cannot set them from the options provided
on the command line until after it has a dtrace_hdl -- but it is dt_vopen()
which provides the dtrace_hdl in the first place!

The only solution to this conundrum is to move both kernpath initialization and
its dependency, module dm_path determination, out of dtrace_update().  The
former can reasonably be moved into dt_kern_path_lookup_by_name(), since that is
the only user of kernpaths once they are initialized: the latter can be set at
module load time, since the module path is needed only for loading things from
the module, and that happens at module load time.

As a side effect this makes DTrace initialization in the trivial no-modules case
twice as fast as it was, as it no longer needs to run a fts() over the kernel
module directory and suck in the names of several hundred modules.