Fix -p and -c options.
This was quite a difficult task, because of the sheer volume of brokenness and
interlocking bugs.

All process manipulations from the process control thread were failing because
ptrace() is thread-specific, and ptrace() control was being taken in the main
dtrace thread, before invocation of the process control thread.  Fixing this
involved migrating the process start / grab machinery into the control thread
and passing it the cmdline and args of the process to start (it already got
the PID of the process to grab, though it used to be the pid of the process
grabbed).

Then there were cleanup bugs, whereby the terminating process sometimes cleaned
up its resources and sometimes did not.  Now, the control thread has
responsibility for cleaning up after itself if and only if it is terminating
because the process it is monitoring has terminated (i.e. an asychronous
termination).  If it is terminating in a situation in which DTrace is waiting
for it (be that due to process invocation/grab problems or due to DTrace
termination), it is responsible only for killing its process and cleaning itself
from the LRU list.  All other data structures (including the process hash) are
untouched.  Further, the entire cmd/dtrace.c process notification handler was
#iffed out, so it didn't notice libdtrace telling it about process termination,
and ran on, dereferencing pointers into freed memory and generally causing
trouble: it also didn't notice attempts to grab or execute nonexistent
processes.  All this cleans up multiple double-frees and other corruption on
normal process termination, which had gone unnoticed becuase of...

... a different process-termination bug: on termination we dropped out of the
state-case in the controlling thread, but did not drop out of the process
monitoring loop, so termination of the process did not cause us to terminate
in turn (as it should have).

Then there was a bug with -p.  When invoking a process with -c, we want to
ptrace() it only until dtrace_continue() is called, then detach from it and let
it keep running without ptrace() interfering with it in all the horrible ways it
does.  When grabbing a process with -p, we cannot get away with that, as we have
no parent/child relationship with the process to tell when it dies: so, in that
situation, we merely PTRACE_CONT it and stay ptrace()-attached.

Then there was a problem with Pgrab() itself, also affecting -p.  We considered
a PT_ATTACH failure to not be an error (as it isn't: if we can't attach, we can
at least observe, even if we can't detect process death), but we skipped
waitpid()ding for the process if PT_ATTACH failed, which meant that the post-
attach stop was reported to Pwait(), which was not expecting a SIGTRAP and
halted the process in protest.

Then there was a race condition on process creation.  New processes used to be
allowed to run freely after startup until halted a short time after startup by a
Pgrab() call.  This meant we lost all possibility of tracing their startup,
which was silly because that was a likely reason to run them under dtrace -c in
the first place.  Now, we use a standard pattern for this sort of thing, halting
them before execve() via a raise(SIGSTOP), flipping on the PTRACE_O_TRACEEXEC
option in the parent, allowing the child to run to execve()-or-failure, then
flipping the option off again, but *not* resuming it until later, when
dtrace_proc_continue() is called.  This allows us to trace absolutely all of
process execution from successful invocation to death.

With all those bugs solved, -c and -p work: we verify this with four new tests
to ensure that we can invoke processes with -c and -p and watch the termination
of the latter and the startup and termination of the former, and that attempts
to invoke or grab nonexistent processes fail and cause dtrace(1) to exit.