Testsuite driver: sorted results comparison.
DTrace output can theoretically appear in arbitrary order due to migration of
tasks between CPUs and the way buffers are linearly accumulated across CPUs
without concern for relative order.  This leads to occasional spurious test
failures, especially under load.  Though output order is guaranteed within
a single D clause, this is not easy to validate: best to assume, for now,
that no ordering constraints exist at all.

Since this behaviour means it is a mistake to rely on test output ordering in
any circumstance, we shouldn't be checking it in expected-result comparison. So
this change switches to comparing sorted results (though the results emitted to
the logfile, and the captured results, are unsorted for ease of reading).

Also, when DTrace hangs, we want to tell the user about it even if the user
typed 'make check', running the testsuite in silent mode.