runtest: Avoid test-output corruption when tests hang.
If a test hangs, we detect and kill it: if it refuses to die, we terminate the
test.  However, if the test dies but a subprocess continues running and
producing output, it can keep dumping output into its stdout or stderr
temporary files, leading to output for subsequent tests that contains the
expected output, then a large hole, then more output from the earlier test
(since we overwrite the temporary file for each new test run rather than
deleting it).

We fix this by explicitly deleting the test output files once we are done with
them, and by using a different temporary file for every test run.  Either of
these should fix the problem, so both at once surely will.