Enable test/unittest/aggs/tst.subr.d.
Now that #inclusion of <sys/dtrace.h> works, we can enable this test, with a
couple of tweaks.

The test is intended to verify that we can at least parse all the D subrs -- but
we can't, since two of them (msgsize() and msgdsize()) will never be implemented
and depend on a nonexistent type, and three (inet_ntoa(), inet_ntoa6(), and
inet_ntop()) depend on a type defined by ip.d, which isn't working yet.

So we have to compensate for this with an ugly count of the number of
not-yet-implemented subrs.