Do not filter cloned functions out of the kernel symbol table.
We filter out all symbols whose names contain a dot from the kernel symbol
table, because the vast majority are of no interest at all to DTrace.

But there is one exception: symbols whose names contain '.clone.' are generated
by GCC's function-cloning optimization, and could very possibly contain probe
points or call functions that themselves contain probe points.  So we should
keep track of their name<->address mapping.

(Transforming the name to not mention the .clone.[number] is possible but should
be done by the infrastructure that actually emits the names: it is retained at
this point in order that the name<->address mapping remains 1:1.)