Provide translators for io, sched, and proc SDT probes, and addtional info.
This commit adds a d_path() subroutine to DIF that converts a 'struct path *'
into a pathname string.  It is used in the io xlator.  It also provides a more
accurate regs.d script, specifying identifiers for processor registers.

The commit also adds the required xlators for io, proc, and sched probes,
making it possible to ensure that the arguments to the SDT probes are
consistent with the DTrace documentation.