Enable tests that use the raise() action.
Now that the DTrace core (at the kernel level) supports the raise() action,
tests that depend o nit can be enabled.  Triggers that were used for these
tests tended to use getpid() in a loop as a syscall to probe for, but Linux
tends to cache the result of a getpid() call and therefore does not actually
execute the syscall at every invocation.  The triggers have been reworked to
use an ioctl() instead.

This commit also enables the io/tst.fds.d test, which exercises both the raise
action, and io xlator for file structures (accessed through the fds global
variable).