testsuite: work on kernels in which ext3 is a non-loaded module.
If the ext3 module is not loaded, unittest/scalars/tst.misc.d fails.
So it should be loaded if possible.  However, if its unloading is impossible
this is not a problem: it is probably in use by an existing filesystem.

So introduce a new flag in test/modules, @@quiet, which indicates that errors
during the removal of the module on this line should be hidden, and add
ext3 to test/modules with this flag attached.