Unknown-type-lookup reduction: translator input.
In translators:

translator foo_ < struct bar * baz > = ...

the 'baz' part must always be an identifier, but id_or_type() didn't realise
that and was doing the usual massive-unknown-type lookup dance.  Now that we
have translators in the default set of dlibs, this must be fixed.

We do this by introducing a new flag in the parser control block to indicate
that we are in between the < > in a translator definition, and teach the grammar
to handle this.  Token lookahead makes this slightly tricky: we have to jump in
before the type name and flip on this flag in order that it is guaranteed to be
on by the time we get to the identifier (we don't care if it is on while the
type name is lexed).  Then, in id_or_type(), we can say that if we are in
state S0 (used not only for program clauses but for the insides of definitions)
and the next non-whitespace token is a >, and we are in a translator input
declaration, then this must be an identifier.