No longer use the running kernel version number in the objdir name.
This was, at best, confusing, since it suggested that uname -r was a useful
thing to do in the build system.  In fact we cannot rely on any useful
properties of the running kernel (e.g. on build bots), so this served only
to mislead.

Now, we use simply 'build'. The rarely-used runtest.sh regtesting infrastructure
still works despite this change : just rename a 'build' directory to anything
starting with 'build' and it will pick it up.