Warn if replacement of tokens in .d.in scripts fails.  Generated sed to objdir.
The existing procedure which preprocesses a C file to produce a
not-entirely-syntactically-valid sed script which is then run over .d.in files
containing replaceable @tokens@ to produce .d files is suboptimal.  Because
the C preprocessor is used, complex sed scripts are impossible: among other
things, you cannot set up a sed script which complains if it is asked to
replace a token @FOO@ with the string FOO, as would happen if one or more of
the #defines mentioned in the headers do not exist.  This is made doubly likely
since the preprocessing step throws away errors (a good few errors are expected,
since we are #including non-unifdeffed kernel header files from outside the
kernel). So we can't even diagnose the case where entire headers don't exist,
which then tends to produce invalid .d files until 'make check' time.

This is not ideal.

So, instead, the .d production process is changed to

.sed.in -> C preprocess -> m4 with sed-replace.m4 -> .sed ->
grep for invalidity tokens -> sed-substitution.

The .sed.in C preprocessing phase still does the same replacement as before, but
rather than converting SED_REPLACE(A) into s/A/expansion-of-A/ it turns it into
an M4 macro call, __sed_replace(A,expansion-of-A).  sed-replace.m4 contains the
definition of this macro, which either emits a sed expression or the static
string @@@@SUBST-FAIL@@@@ if the preprocessor didn't find the C macro we're
trying to replace.  The Build script can then grep for this symbol and fail
if it is found.  (We also let M4 do the " -> @ translation which used to be done
by a tr call in the Build script.))

As a bonus, this lets us throw away the /* ... */ comments, and the expansion of
the C header files, at M4 time: these are invalid sed, do not belong in a sed
script, and needed tr transformation to prevent sed from interpreting & in C
code as a backreference.

This does mean that the resulting .sed files have no comments at all: that's
OK since they are build artifacts.  As build artifacts they belong in the
objdir, so we migrate them in there (using GNU make secondary expansion to
determine their location in pattern rules).

Finally, a side-effect of this is that sedding of D libraries not in
WORKING_DLIBS is suppressed, since by their very nature these libraries are
not ported yet and almost certainly contain non-substituted preprocessor tokens.