Specfile: bound the devel kernel requirement at one end only.
RPM does not support version ranges. We need a better workaround later, but for
now, simply constrain the devel kernel at the lower end and restrict it to
DTrace kenrels.  (This will fail if a >.201 DTrace kernel comes out and
we do not rerelease userspace or fix this, but this is unlikely for now.)