Specfile: Bind build requirements tightly to the current dtrace kernel.
My earlier attempt to work with whatever devel kernels may be present on the
system is both complex and runs foul of RPM's inability to specify closed ranges
in requirements.  Furthermore, it is useless: this is only a build requirement,
not a runtime requirement, so tightly binding to a single devel kernel will not
impose on DTrace users.

So just do that.  Whenever releasing a new dtrace-utils, it is probably a good
idea to boost %kver in the specfile (but unless kernel headers change in ways
they rarely do, not doing so will do no harm.)