DTrace-independent CTF.
Catering for the renaming of dtrace_ctf.ko to ctf.ko is easy enough: we just
need to change the name of the module we look up.  But the CTF loading code
assumes that when it finds CTF with a parent name, it can load whatever module
has that name and extract CTF from it.  This is a nice property worth retaining,
but runs into the problem that the CTF parent name is now "shared_ctf", while
the kernel module it is part of is "ctf.ko" -- not the same.  We fix this by
forcibly changing the module name of ctf.ko to 'shared_ctf' at dt_module load
time, and introducing a new DT_DM #define, DT_DM_SHARED, to indicate that the
module with this flag is the shared CTF module and that its section name should
be constructed specially.

(Retaining the "CTF parent names are always the name of another module" property
will let us potentially set up hierarchies of nested CTF if we want to
parallelize CTF file construction in the future, so this few lines of extra
complexity is worth it.)