Fix notypeload failure on rare types that exist only in dtrace.ko.
Some rare types are used only in a few modules, dtrace.ko and translators.  Right
now, if .config is so configured that none of those modules are built in, DTrace
must iterate across numerous modules to find the type in dtrace.ko, costing
startup time.  Since dtrace userspace and dtrace.ko are fairly tightly tied,
it is reasonable for translators to reference types that exist only there
without incurring extra cost at every startup: so shuffle the dtrrace module
to the start of the module list, making type lookups from there cheap.

Also, remove shared_ctf from the shuffle-to-the-front: it is never looked up
directly, but always via parent traversals.  So type lookups in shared_ctf
are a waste of time because if a type could be found in there, the lookup
in vmlinux (also at the front of the module list) would always have succeeded.
The waste of time is only tiny (the high-cost part of type lookup is the first
lookup, when the module is loaded, and shared_ctf is always loaded).  But
nonetheless, let's avoid incurring it unnecessarily.

Also, fix tst.notypeload.sh so that it can handle either lookups in either two
or three modules (possibly including dtrace.ko), and adjust it to emit the list
of modules CTF lookups were done in when it fails.